Cleanup some documentation on the dkron package
This should help clarify some on the purpose of exported functions,
constants, errors, types, and interfaces. I changed some functions
to be unexported because they should not be modified outside the
dkron package. I hope to clean up some more in the future when
I can find some more time.