Fix race condition in job update
As reported https://github.com/victorcoder/dkron/issues/299

This fixes a race condition with the job update after execution.

In this PR the lock is refactored to an Atomic job put, that reads the job, updates it and performs a CAS operation to update it in the store.

If this update fails, it retryies it until it's updated.

This ensures no race condition between nodes and efectively fixes the reported error.