Fix windows plugins for v0.10
These plugins were loaded off disk with an .exe
extension on windows. This was making its way into
the executor map and therefore providing a different
experience on windows vs unix/linux.

Also corrects the error message that would be logged when
this effects a user. The executor will be nil when the log
entry is written.