Filter jobs by tags
From now on we can assign labels to the jobs. And query (`GET /jobs`) them with labels filter. This change could be useful for the users who have to build UI and display background jobs on their own. One of the use case when you have a different kind of jobs like the system jobs and some application related and you don't want the app to see the jobs not assigned to it.