Change the Job running mutex method
The former mutex was causing contention in low schedule time Jobs, and causing other weird behaviours on deletes and updates due to that contention.

Also running above serf broadcast rate was causing queue length contention in serf itself.

Replacing it for a simple boolean check and skipping runs under 3 something seconds, that is the minimum broadcast time, allows to notify the user about the fact the execution skipped.