Add prometheus endpoint for metrics
The existing metrics package used in dkron already supports prometheus, so this just adds in the necessary pieces to hook it up to the API. There is a config value to enable serving this endpoint to allow users the choice with a default of false to keep today's functionality. I know that it was discussed in #494 that the metrics system was likely going to be updated, but this seems like an easy step in the meantime.

The documentation for metrics was updated to align with current configuration keys.