refactor: Use gRPC streaming calls to run jobs
Serf queries works over UDP sending messages to nodes and expecting a
response, there is no guarantee or verification of the message delivery.
This was causing some job executions to be lost, around 1%-2% on edge
cases.

Solution

Agents will listen to gRPC calls from the servers, using server-side
streaming, the other way around as currently, agents will send Execution
progress as now.

Servers will actively order nodes to run the jobs, so we can verify the
job is actually being executed or report an error in case it's not.

This maintains the same guarantees in job status reporting and
streaming, also adding the possibility to cancel an execution in the
future.