React admin Web UI (#864)
React Admin UI

Rewrite WebUI using React with Reac admin, using Typescript to match
project codebase, style and take advantage of its features.

Modification to the API are necessary to correctly use the features of
RA Datagrid, pager, etc.

It stores Jobs in JSON to BuntDB, but if existing values are present in
Proto format it converts them, this was necessary to use Sorting
capabilities using BundDB indexes.

The new UI is optional at the moment, it is possible to fallback to the
old ui by setting the --ui param to false